move account details section component to a separate file
